+++ peruvian-game-server/helpers/Common.js
@@ -1,0 +1,3 @@
+exports.getRandInt = (max) => Math.floor(Math.random() * max);
+
+exports.getObjValues = (obj) => Object.keys(obj).map((key) => obj[key]);
