--- peruvian-game-server/questions.json
@@ -1,109 +1,0 @@
-{
-	"nutrition": [
-		{
-			"id": "5393b83f-1d78-4e42-961c-b419872d2814",
-			"question": "What is an apple?",
-			"type": "multiple-choice",
-			"difficulty": 1,
-			"answer": "A fruit,",
-			"options": [ "A fruit", "an orange" ]
-		},
-		{
-			"id": "10862e1d-442c-418a-903d-d8435cb2f6ac",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "6ee2b7e0-f283-42b8-a94f-96e79c5aab0b",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "fdd06a20-3fc2-49e4-b21b-aa5b36066414",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "149fb705-34a7-454d-8f73-573a04a0387d",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "6a25ca2a-3d63-471d-89ba-49e0556ea95c",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "ff3b6a02-e3f9-4d33-b6cf-62a788b3d84c",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "5c63f5cf-247a-4b8d-b0a4-e7235b376568",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "454d5245-f621-42b6-803a-954938d7859e",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		},
-		{
-			"id": "6106636b-bf54-4d3c-8678-0181f032eff1",
-			"question": "Is a mango a fruit or vegetable?",
-			"type": "text",
-			"difficulty": 1,
-			"answer": "fruit"
-		}
-	],
-	"uuids": [
-		"USE THESE FOR THE IDS IF YOU NEED THEM",
-
-		"153e1674-88b9-4314-b5e9-e68ca0565765",
-		"eafed659-289a-4ba4-8160-9f1a26eb58f4",
-		"5f9c3625-22c0-4feb-a396-380e6026dd6e",
-		"6c0ec66e-ff8c-486b-b276-b46f50fe6a00",
-		"37564bcd-2847-4798-a2c9-9228cd1f047b",
-		"f6ebf38d-12f3-4dcd-ac10-dc355852f891",
-		"fb69bb31-067c-4f82-a46b-bbfe1a6b70fa",
-		"0a0b50c5-0956-4304-ab77-132ceeb976c4",
-		"36d98a34-72e4-47d8-bb1f-d43e1f389ecb",
-		"c4277763-4e0e-49e2-a121-c76ea37f9fd5",
-		"a299fef6-8c78-4897-934c-f184940d2b43",
-		"56ef2b32-f9f2-4a47-bffe-df4054bd98fe",
-		"47c16a2e-8dd7-4c62-b893-d11be3dda163",
-		"43d12509-78c0-4298-9bce-c8cc1a81e170",
-		"72ec3187-af7f-44d3-9bee-b5787ee8260a",
-		"a26d6515-1167-449a-bc63-0f4741cbd20c",
-		"36bfa2dd-23d8-4659-ac04-ec3714ec264d",
-		"aa8e3492-b5e5-45c6-bd19-84f6aacf339d",
-		"808dc152-e716-4116-8f5e-5632b602fa97",
-		"509a61a9-f967-4f3e-9d08-e9cd5139dd44",
-		"aa3f4b2e-4218-4b97-893b-a45be34241f3",
-		"dfc4c94f-fc7a-4720-8671-1c2f1e3f345e",
-		"481da28b-e42a-45b7-b3be-b18df8f4a959",
-		"3f00a22d-2c11-4bc7-bbdb-4e14e9fb3a15",
-		"d511c3a0-0bb9-4717-be57-aa21ce9db61a",
-		"daa58ba0-e703-49a6-8be5-8684b140d6b5",
-		"4bb24048-5295-48ba-9c0e-2e455352a9e4",
-		"6388b03b-eed5-45c2-beca-49b6f829211d",
-		"0850fc2e-4bf1-49cf-8f5d-8e693e404c0e",
-		"52db621f-c886-4416-8e6d-18b6efe2cf4d"
-	]
-}
